GO-3192 Move check to table pkg
